Code refactor and tidy
